+++ engine/common/rpc/convert/transactions_test.go
@@ -1,0 +1,22 @@
+package convert_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"github.com/onflow/flow-go/engine/common/rpc/convert"
+	"github.com/onflow/flow-go/model/flow"
+	"github.com/onflow/flow-go/utils/unittest"
+)
+
+func TestConvertTransaction(t *testing.T) {
+	tx := unittest.TransactionBodyFixture()
+
+	msg := convert.TransactionToMessage(tx)
+	converted, err := convert.MessageToTransaction(msg, flow.Testnet.Chain())
+	assert.Nil(t, err)
+
+	assert.Equal(t, tx, converted)
+	assert.Equal(t, tx.ID(), converted.ID())
+}
